Add files via upload
Storing the codes in .txt format has several benefits. First, it allows for easy access and readability of the codes, even without specialized software. The code can be opened with any text editor, and changes can be made as needed. Additionally, .txt files are lightweight and do not require any special software to create or modify them. They can be stored and transferred easily, without concerns about compatibility issues or software version changes. This makes them a reliable and long-term storage format for the code, which can be accessed and modified at any time, regardless of any updates or changes to the libraries or software used. Overall, storing codes in .txt format provides a simple and versatile way to manage and preserve the code over time.